rename client folder, to-do: unit test coverage
