+++ proxy_agent_shared/src/lib.rs
@@ -2,11 +2,11 @@
 // SPDX-License-Identifier: MIT
 
 pub mod certificate;
-pub mod client;
 pub mod common;
 pub mod error;
 #[cfg(windows)]
 pub mod etw;
+pub mod host_clients;
 pub mod logger;
 pub mod misc_helpers;
 pub mod proxy_agent_aggregate_status;
